\name{fizz_buzz}
\alias{fizz_buzz}
\title{Fizz Buzz Game}
\usage{
fizz_buzz(n)
}
\arguments{
\item{n}{An integer specifying the number of terms to generate.}
}
\value{
A vector containing the first n terms of the Fizz Buzz game.
}
\description{
This function generates the first n terms of the Fizz Buzz game.
}
\examples{
fizz_buzz(15)

}
