% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of daily precipitation and snow levels along with maximum and minimum temperatures in Flagstaff, AZ.}
\format{
a data frame with 25449 observations
\describe{
  \item{DATE}{Date.}
  \item{PRCP}{Precipitation.}
  \item{SNOW}{Snowfall.}
  \item{tmax}{Maximum temperature.}
  \item{tmin}{Minimum temperature.}
}
}
\source{
\url{https://www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of daily precipitation and snow levels along with maximum and minimum temperatures in Flagstaff, AZ.
}
\keyword{datasets}
